spark-server: keep the device-arch gate off preflight.rs's line budget
`preflight.rs` sat at 499 LoC on main, so the two lines the arch gate needs
would have pushed it over the 500-line cap. Move both `init_gpu_backend` arms
into `preflight/gpu_backend.rs` — a self-contained unit (construct the
backend, latch the free-memory baseline, log the device) that reads better
beside its Metal twin than buried between the reserve and post-load audits —
and give `kernel_gate` a `gate_device_arch` wrapper so the call site is one
line. No behaviour change: same order, same errors, same gate.
