+++ crates/spark-server/src/main_modules/serve_phases/preflight.rs
@@ -2,14 +2,17 @@
 
 //! GPU init + pre-load reserve preflight + post-load OOM check.
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 
 use atlas_core::config::ModelConfig;
 
 use crate::cli;
 
+mod gpu_backend;
 mod per_sequence_state;
 mod ssm_h_fp16;
+#[cfg(any(feature = "cuda", feature = "metal"))]
+pub(crate) use gpu_backend::init_gpu_backend;
 use {per_sequence_state::per_sequence_reserve, ssm_h_fp16::ssm_h_fp16_preconditions};
 
 pub(crate) struct ReservePreflight {
@@ -330,81 +333,6 @@ pub(crate) fn preflight_reserve(
         ssm_prefill_chunk,
         max_batch_tokens_pre,
     })
-}
-
-/// Initialize the GPU backend for the active feature.
-///
-/// Compile-time dispatch:
-/// - `cuda` feature → `AtlasCudaBackend` loading PTX modules from `ptx_set`.
-/// - `metal` feature → `MetalGpuBackend` loading metallib modules from
-///   `ptx_set` as well. Both arms register the RESOLVED target's modules;
-///   `metallib_modules()` is a plain alias of target 0, so registering from
-///   it served another model's kernels in a multi-target build.
-#[cfg(feature = "cuda")]
-pub(crate) fn init_gpu_backend(
-    args: &cli::ServeArgs,
-    ptx_set: &atlas_kernels::TargetPtxSet,
-) -> Result<(Box<dyn spark_runtime::gpu::GpuBackend>, usize)> {
-    // BEFORE the backend, because constructing it loads every PTX module and
-    // a driver-side arch rejection names neither the arch nor the GPU. The
-    // compiled arch is only knowable here: `AtlasCudaBackend::new` takes module
-    // blobs. WHICH arch string to hand over is `preflight_arch`'s decision and
-    // not this call site's — `ptx_set.target.arch` is the base SM with the
-    // feature suffix stripped, and judging that would wave `sm_90a` kernels
-    // onto a CC 10.0 device.
-    super::kernel_gate::gate_arch_preflight(
-        args.check_kernels,
-        ptx_set,
-        spark_runtime::cuda_backend::arch_preflight::preflight_device_arch(
-            args.gpu_ordinal,
-            spark_runtime::cuda_backend::arch_preflight::preflight_arch(ptx_set),
-        ),
-    )?;
-
-    let backend =
-        spark_runtime::cuda_backend::AtlasCudaBackend::new(args.gpu_ordinal, &ptx_set.modules)
-            .context("Failed to initialize CUDA backend")?;
-
-    let gpu: Box<dyn spark_runtime::gpu::GpuBackend> = Box::new(backend);
-    let total_mem = gpu.total_memory()?;
-    let free_mem = gpu.free_memory()?;
-    // Baseline for self-relative KV budgeting: free memory now (post context +
-    // PTX modules, pre weights) minus free-at-build = this process's own
-    // footprint, co-tenants excluded. See gpu::baseline_free_bytes.
-    spark_runtime::gpu::set_baseline_free_bytes(free_mem);
-    tracing::info!(
-        "GPU {}: {:.1} GB total, {:.1} GB free",
-        args.gpu_ordinal,
-        total_mem as f64 / (1024.0 * 1024.0 * 1024.0),
-        free_mem as f64 / (1024.0 * 1024.0 * 1024.0),
-    );
-    Ok((gpu, free_mem))
-}
-
-#[cfg(all(feature = "metal", not(feature = "cuda")))]
-pub(crate) fn init_gpu_backend(
-    args: &cli::ServeArgs,
-    ptx_set: &atlas_kernels::TargetPtxSet,
-) -> Result<(Box<dyn spark_runtime::gpu::GpuBackend>, usize)> {
-    // The RESOLVED target's modules, exactly like the CUDA arm above.
-    // `metallib_modules()` is an alias of `ptx_modules()`, which build-codegen
-    // emits as a plain alias of TARGET 0 in a multi-target build — so this
-    // registered another model's kernels and every lookup for the model
-    // actually being served failed.
-    let gpu: Box<dyn spark_runtime::gpu::GpuBackend> = Box::new(
-        spark_runtime::metal_backend::MetalGpuBackend::new(args.gpu_ordinal, &ptx_set.modules)
-            .context("Failed to initialize Metal backend")?,
-    );
-    let total_mem = gpu.total_memory()?;
-    let free_mem = gpu.free_memory()?;
-    spark_runtime::gpu::set_baseline_free_bytes(free_mem);
-    tracing::info!(
-        "Metal device {}: {:.1} GB total, {:.1} GB free",
-        args.gpu_ordinal,
-        total_mem as f64 / (1024.0 * 1024.0 * 1024.0),
-        free_mem as f64 / (1024.0 * 1024.0 * 1024.0),
-    );
-    Ok((gpu, free_mem))
 }
 
 #[allow(clippy::too_many_arguments)]
